pyproject.toml
[tool.poetry]
name = "dndserver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.0"
pyaml-env = "^1.2.1"
twisted = {extras = ["tls"], version = "^22.10.0"}
protobuf = "^4.22.3"
dataset = "^1.6.0"
sqlalchemy-utils = "^0.41.0"
argon2-cffi = "^21.3.0"
arrow = "^1.2.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120